requirements.txt
# requirements.txt
# Bibliothèques communes pour tout le projet

# Base de données
psycopg2-binary==2.9.9
pgvector==0.2.5

# Machine Learning et NLP
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Utilitaires
python-dotenv==1.0.0